supercollider and microtonal support

+++ comm/senderdebug.cpp
@@ -34,7 +34,7 @@ void SenderDebug::pc(int c, int v1)
     qDebug() << "SenderDebug::pc(" << c << "," << v1 << ")";
 }
 
-void SenderDebug::note(int c, int voiceId, int v1, int v2)
+void SenderDebug::note(int c, int voiceId, double v1, int v2)
 {
     qDebug() << "SenderDebug::note(" << c << "," << voiceId << "," << v1 << "," << v2 << ")";
 }
