+++ comm/senderoscpuredata.h
@@ -29,20 +29,17 @@ public:
     SenderOscPuredata(RC1 * rc1);
     virtual void cc(int c, int voiceId, int cc, double v1);
     virtual void pc(int c, int v1);
-    virtual void note(int c, int voiceId, int f, int vel);
+    virtual void note(int c, int voiceId, double f, int vel);
     virtual void setDestination(QHostAddress a,int p);
+
 private:
     QOscClientInterface* oscout;
     RC1 * rc1;
-
     int prog;           // current program
     quint8 * notestate;      // currently played notes
     int * ccstate;       // current ccval;
     quint8 onNoteCnt;   // count of on Notes
-
-
     void sendOsc(QString path, QVariant list);
-
 };
 
 #endif // SENDEROSCPUREDATA_H
